+++ src/check_jsonschema/loaders/schema.py
@@ -9,7 +9,7 @@ import jsonschema
 import ruamel.yaml
 
 from ..builtin_schemas import get_builtin_schema
-from ..cachedownloader import CacheDownloader, FailedDownloadError
+from ..cachedownloader import CacheDownloader
 from ..formats import FormatOptions, make_format_checker
 from ..utils import is_url_ish
 from .errors import SchemaParseError, UnsupportedUrlScheme
@@ -72,10 +72,8 @@ class HttpSchemaReader:
         url,
         cache_filename: t.Optional[str],
         disable_cache: bool,
-        failover_builtin_schema: t.Optional[str] = None,
     ):
         self.url = url
-        self.failover_builtin_schema = failover_builtin_schema
         self.downloader = CacheDownloader(
             url, cache_filename, disable_cache=disable_cache
         )
@@ -84,19 +82,8 @@ class HttpSchemaReader:
         return self.url
 
     def read_schema(self):
-        try:
-            with self.downloader.open() as fp:
-                return _json_load_schema(self.url, fp)
-        except FailedDownloadError as err:
-            if self.failover_builtin_schema:
-                val = get_builtin_schema(self.failover_builtin_schema)
-                if val:
-                    return val
-                else:
-                    raise ValueError(
-                        f"failover schema {self.failover_builtin_schema} not valid"
-                    ) from err
-            raise
+        with self.downloader.open() as fp:
+            return _json_load_schema(self.url, fp)
 
 
 class SchemaLoader:
@@ -105,13 +92,11 @@ class SchemaLoader:
         schemafile: str,
         cache_filename: t.Optional[str] = None,
         disable_cache: bool = False,
-        failover_builtin_schema: t.Optional[str] = None,
     ):
         # record input parameters (these are not to be modified)
         self.schemafile = schemafile
         self.cache_filename = cache_filename
         self.disable_cache = disable_cache
-        self.failover_builtin_schema = failover_builtin_schema
 
         # if the schema location is a URL, which may include a file:// URL, parse it
         self.url_info = None
@@ -130,7 +115,6 @@ class SchemaLoader:
                 self.schemafile,
                 self.cache_filename,
                 self.disable_cache,
-                self.failover_builtin_schema,
             )
         elif self.url_info.scheme in ("file", ""):
             return LocalSchemaReader(self.url_info.path)
@@ -189,7 +173,6 @@ def schema_loader_from_args(args) -> SchemaLoader:
             args.schemafile,
             args.cache_filename,
             args.no_cache,
-            args.failover_builtin_schema,
         )
     else:
         return BuiltinSchemaLoader(args.builtin_schema)