+++ vendor-schemas.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python3
+from __future__ import annotations
+
 import datetime
 import hashlib
+import os
 
 import requests
 
@@ -9,17 +12,19 @@ from check_jsonschema.catalog import SCHEMA_CATALOG
 today = datetime.datetime.today().strftime("%Y-%m-%d")
 
 
-def download_schema(schema_name, schema_url) -> bool:
+def download_schema(schema_name: str, schema_url: str) -> bool:
     print(f"downloading {schema_name} schema to check ({schema_url})")
     res = requests.get(schema_url)
     sha = hashlib.sha256()
     sha.update(res.content)
     new_digest = sha.hexdigest()
 
-    with open(
-        f"src/check_jsonschema/builtin_schemas/vendor/{schema_name}.sha256", "r"
-    ) as fp:
-        prev_digest = fp.read().strip()
+    hashfile = f"src/check_jsonschema/builtin_schemas/vendor/{schema_name}.sha256"
+    if os.path.exists(hashfile):
+        with open(hashfile, "r") as fp:
+            prev_digest: str | None = fp.read().strip()
+    else:
+        prev_digest = None
 
     if new_digest == prev_digest:
         return False
@@ -36,7 +41,7 @@ def download_schema(schema_name, schema_url) -> bool:
     return True
 
 
-def update_changelog():
+def update_changelog() -> None:
     with open("CHANGELOG.md", encoding="utf-8") as fp:
         content = fp.read()
     content = content.replace(
@@ -52,7 +57,7 @@ def update_changelog():
         fp.write(content)
 
 
-def main():
+def main() -> None:
     made_changes = False
     for name, config in SCHEMA_CATALOG.items():
         made_changes = download_schema(name, config["url"]) or made_changes